profile update

+++ app/Http/Controllers/API/UserController.php
@@ -71,6 +71,40 @@ class UserController extends Controller
         return auth('api')->user();
     }
 
+    public function updateProfile(Request $request)
+    {
+        $user = auth('api')->user();
+
+        $this->validate($request, [
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users,email,' . $user->id,
+            'password' => 'sometimes|required|min:6',
+        ]);
+
+        //if new photo uploaded is not equal to current photo
+        if ($request->photo != $user->photo) {
+            $extension = explode('/', mime_content_type($request->photo))[1];
+            $name = time() . '.' . $extension;
+            \Image::make($request->photo)->save(public_path('img/profile/') . $name);
+
+            $request->merge(['photo' => $name]);
+
+            $user_photo = public_path('img/profile/') . $user->photo;
+
+            // delete current photo
+            if (file_exists($user_photo)) {
+                @unlink($user_photo);
+            }
+        }
+
+        if (!empty($request->password)) {
+            $request->merge(['password' => Hash::make($request['password'])]);
+        }
+
+        $user->update($request->all());
+        return ['message' => 'Success'];
+    }
+
     /**
      * Update the specified resource in storage.
      *
@@ -100,6 +134,7 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
         $user->delete();
         return ['message' => 'User Deleted'];
